moved get sprint period to frontend
